Move utilitites to separate modules, create a constant for 255
