+++ src/constants/numbers.ts
@@ -1,3 +1,4 @@
 export const MAX_DECIMALS = 6;
 export const DEFAULT_BLEND_STEPS = 5;
 export const DEFAULT_SHADES_TINTS_STEPS = 5;
+export const BASE_255 = 255;
